Resolved Comments
wrapped error in fmt
added license header to follow OSS Compliance
