+++ sources/dynamodb/streaming.go
@@ -1,3 +1,16 @@
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 package dynamodb
 
 import (
@@ -10,7 +23,7 @@ import (
 
 // NewDynamoDBStream initializes a new DynamoDB Stream for a table with NEW_AND_OLD_IMAGES
 // StreamViewType. If there exists a stream for a given table then it must be of type
-// NEW_IMAGE or NEW_AND_OLD_IMAGES otherwise this table is dropped from streaming migration.
+// NEW_IMAGE or NEW_AND_OLD_IMAGES otherwise streaming changes for this table won't be captured.
 // It returns latest Stream Arn for the table along with any error if encountered.
 func NewDynamoDBStream(client dynamodbiface.DynamoDBAPI, srcTable string) (string, error) {
 	describeTableInput := &dynamodb.DescribeTableInput{
@@ -18,7 +31,7 @@ func NewDynamoDBStream(client dynamodbiface.DynamoDBAPI, srcTable string) (strin
 	}
 	result, err := client.DescribeTable(describeTableInput)
 	if err != nil {
-		return "", err
+		return "", fmt.Errorf("unexpected call to DescribeTable: %v", err)
 	}
 	if result.Table.StreamSpecification != nil {
 		switch *result.Table.StreamSpecification.StreamViewType {
@@ -40,7 +53,7 @@ func NewDynamoDBStream(client dynamodbiface.DynamoDBAPI, srcTable string) (strin
 		}
 		res, err := client.UpdateTable(updateTableInput)
 		if err != nil {
-			return "", err
+			return "", fmt.Errorf("unexpected call to UpdateTable: %v", err)
 		}
 		return *res.TableDescription.LatestStreamArn, nil
 	}
